Folded all `table_constraints/*.rs` submodules into `table_constraints.rs` module
